% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnad_covid.R
\encoding{UTF-8}
\name{pnad_covid_microdados}
\alias{pnad_covid_microdados}
\title{Carregando microdados da PNAD COVID19}
\usage{
pnad_covid_microdados(diretorio_dados, ...)
}
\arguments{
\item{diretorio_dados}{Diretório onde os microdados originais em formato de texto estão armazenados}

\item{...}{vetores com datas das pesquisas de interesse no  formato \code{c('m?s', 'ano')}}
}
\value{
Lista de dataframes, sendo cada entrada um m?s/ano em \code{...}
}
\description{
Extrai e constrói bases de dados da PNAD COVID19 em formato R (.RData) a partir
dos microdados originais, os quais  não são disponibilizados pelo Portal do IBGE
(para informa??es sobre como obter os arquivos originais de dados,
consulte o site do IBGE (www.ibge.gov.br).
Como a pesquisa ainda é publicada pelo IBGE, este programa está em constante atualização.
A função gera uma base para cada mês. Se for o caso, utilize o
comando  \code{\link{bind_rows}} para empilhar as bases.
}
\examples{
datazoom_pnadc('./Desktop', c(5, 2020))

}
